+++ python/cudf/bindings/binops.pyx
@@ -13,6 +13,8 @@ from cudf.dataframe.column import Column
 from libcpp.vector cimport vector
 from libc.stdlib cimport free
 
+from libcpp.string cimport string
+
 from librmm_cffi import librmm as rmm
 
 _BINARY_OP = {
@@ -49,6 +51,23 @@ cdef apply_op_v_v(gdf_column* c_lhs, gdf_column* c_rhs, gdf_column* c_out, op):
             <gdf_column*>c_lhs,
             <gdf_column*>c_rhs,
             c_op)
+
+    cdef int nullct = c_out[0].null_count
+
+    return nullct
+
+cdef apply_op_v_v_udf(gdf_column* c_lhs, gdf_column* c_rhs, gdf_column* c_out, ptx):
+    """
+    Call gdf binary ops between two columns using user-defined function (UDF) defined in "ptx".
+    """
+
+    cdef string cpp_str = ptx.encode('UTF-8')
+    with nogil:
+        binary_operation(
+            <gdf_column*>c_out,
+            <gdf_column*>c_lhs,
+            <gdf_column*>c_rhs,
+            cpp_str)
 
     cdef int nullct = c_out[0].null_count
 
@@ -139,6 +158,33 @@ def apply_op(lhs, rhs, out, op):
     free(c_lhs)
     free(c_rhs)
     free(c_scalar)
+    free(c_out)
+
+    return nullct
+  
+def apply_op_udf(lhs, rhs, out, ptx):
+    """
+    Dispatches a binary op call to the appropriate libcudf function:
+    """
+    check_gdf_compatibility(out)
+    cdef gdf_column* c_lhs = NULL
+    cdef gdf_column* c_rhs = NULL
+    cdef gdf_column* c_out = column_view_from_column(out)
+
+    check_gdf_compatibility(lhs)
+    check_gdf_compatibility(rhs)
+    c_lhs = column_view_from_column(lhs)
+    c_rhs = column_view_from_column(rhs)
+
+    nullct = apply_op_v_v_udf(
+        <gdf_column*>c_lhs,
+        <gdf_column*>c_rhs,
+        <gdf_column*>c_out,
+        ptx
+    )
+
+    free(c_lhs)
+    free(c_rhs)
     free(c_out)
 
     return nullct