+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - PR #1863 Series and Dataframe methods for all and any
 - PR #1921 Add additional formats for typecasting to/from strings
 - PR #1807 Add Series.dropna()
+- PR #1987 Allow user defined functions in the form of ptx code to be passed to binops
 - PR #1948 Add operator functions like `Series.add()` to DataFrame and Series
 - PR #1954 Add skip test argument to GPU build script
 - PR #1984 Add rolling window operations Series.rolling() and DataFrame.rolling()